YUK-15023 CBC Analog point data is being truncated
An auxiliary function had a long for an incoming value parameter, that was being passed through to a function that expected doubles, truncating the data.

+++ yukon-server/CapControl/cctwowaycbcpoints.cpp
@@ -341,7 +341,7 @@ bool CtiCCTwoWayPoints::isTimestampNew(long pointID, CtiTime timestamp)
     return ! (_pointValues.getPointTime(pointID, prevTime) && timestamp <= prevTime);
 }
 
-bool CtiCCTwoWayPoints::setTwoWayPointValue(long pointID, long value, CtiPointType_t type, CtiTime timestamp)
+bool CtiCCTwoWayPoints::setTwoWayPointValue(long pointID, double value, CtiPointType_t type, CtiTime timestamp)
 {
     boost::optional<CtiPointType_t> lookupType = Cti::mapFind( _pointidPointtypeMap, pointID );
     if ( lookupType && *lookupType == type && isTimestampNew( pointID, timestamp ) )
